+++ spec/spec_helper.rb
@@ -1,16 +1,22 @@
 # frozen_string_literal: true
 
-require "x402/rails"
-require 'simplecov'
-require 'simplecov-formatter-badge'
-
+require "simplecov"
+require "simplecov-formatter-badge"
 
 SimpleCov.start do
-  SimpleCov.formatters = SimpleCov::Formatter::MultiFormatter.new([
+  formatter SimpleCov::Formatter::MultiFormatter.new([
     SimpleCov::Formatter::HTMLFormatter,
     SimpleCov::Formatter::BadgeFormatter
   ])
+  add_filter "/spec/"
 end
+
+require "rails"
+require "x402/rails"
+
+Dir[File.expand_path("support/**/*.rb", __dir__)].sort.each { |f| require f }
+
+Rails.logger = Logger.new(File::NULL)
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -22,4 +28,7 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.order = :random
+  Kernel.srand config.seed
 end